# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: tff order stats API
  description: |
    External REST API for tff order stats.

    This API allows external systems to create, read, and update Tesla orders.
    All endpoints require API key authentication via the `X-API-Key` header.

    ## Authentication
    Include your API key in the `X-API-Key` header with every request:
    ```
    X-API-Key: your-api-key-here
    ```

    ## Response Format
    All responses follow a consistent envelope format:
    ```json
    {
      "success": true,
      "data": { ... },
      "meta": {
        "timestamp": "2025-02-12T10:30:00Z",
        "version": "v1"
      }
    }
    ```

    ## Date Format
    All dates use German format: `DD.MM.YYYY` (e.g., `15.01.2025`)
  version: 1.0.0
  contact:
    name: tff order stats
servers:
  - url: 'https://tff-order-stats.de/api/v1'
    description: Production
  - url: 'http://localhost:3000/api/v1'
    description: Local Development
security:
  - ApiKeyAuth: []
...
paths:
  /orders:
    get:
      tags:
        - Orders
      summary: List all orders
      description: Returns a paginated list of orders with optional filtering
      operationId: listOrders
      parameters:
        - name: limit
          in: query
          description: Maximum number of orders to return (1-100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          description: Number of orders to skip
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: country
          in: query
          description: Filter by country code (e.g., de, at, ch)
          schema:
            type: string
        - name: model
          in: query
          description: Filter by model (standard, premium, performance)
          schema:
            type: string
            enum: [standard, premium, performance]
        - name: archived
          in: query
          description: Include archived orders
          schema:
            type: boolean
            default: false
        - name: vehicleType
          in: query
          description: Filter by vehicle type
          schema:
            type: string
            enum: [Model Y, Model 3]
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Order'
                      meta:
                        allOf:
                          - $ref: '#/components/schemas/Meta'
                          - type: object
                            properties:
                              pagination:
                                $ref: '#/components/schemas/Pagination'
              example:
                success: true
                data:
                  - id: "clx123abc"
                    name: "MaxMustermann"
                    vehicleType: "Model Y"
                    orderDate: "15.01.2025"
                    country: "de"
                    model: "performance"
                    range: null
                    drive: "awd"
                    color: "ultra_red"
                    interior: "black"
                    wheels: "21"
                    towHitch: "ja"
                    autopilot: "fsd"
                    seats: "5"
                    source: null
                    tostUserId: null
                    deliveryWindow: "Februar 2025"
                    deliveryLocation: "München"
                    vin: null
                    vinReceivedDate: null
                    papersReceivedDate: null
                    productionDate: null
                    typeApproval: null
                    typeVariant: null
                    deliveryDate: null
                    orderToProduction: null
                    orderToVin: null
                    orderToDelivery: null
                    orderToPapers: null
                    papersToDelivery: null
                    archived: false
                    archivedAt: null
                    createdAt: "2025-02-12T10:30:00.000Z"
                    updatedAt: "2025-02-12T10:30:00.000Z"
                meta:
                  timestamp: "2025-02-12T10:30:00.000Z"
                  version: "v1"
                  pagination:
                    total: 150
                    limit: 50
                    offset: 0
                    hasMore: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'

    post:
      tags:
        - Orders
      summary: Create a new order
      description: Creates a new order with the provided details
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
            example:
              name: "MaxMustermann"
              vehicleType: "Model Y"
              orderDate: "15.01.2025"
              country: "de"
              model: "performance"
              drive: "awd"
              color: "ultra_red"
              interior: "black"
              wheels: "21"
              towHitch: "ja"
              autopilot: "fsd"
              seats: "5"
              deliveryWindow: "Februar 2025"
              deliveryLocation: "München"
              editCode: "MySecure123"
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CreateOrderResponse'
              example:
                success: true
                data:
                  id: "clx789ghi"
                  editCode: "MySecure123"
                  message: "Order created successfully"
                meta:
                  timestamp: "2025-02-12T10:30:00.000Z"
                  version: "v1"
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for general API authentication
    TostApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Dedicated API key for TOST integration (separate from general API key)
  schemas:
    Order:
      type: object
      properties:
        id:
          type: string
          description: Unique order identifier
        name:
          type: string
          description: Customer username
        vehicleType:
          type: string
          nullable: true
          enum: [Model Y, Model 3]
          description: Vehicle type (Model Y or Model 3)
        orderDate:
          type: string
          nullable: true
          description: Order date (DD.MM.YYYY format)
        country:
          type: string
          nullable: true
          description: Country code
        model:
          type: string
          nullable: true
          description: |
            Tesla model trim. Values depend on vehicleType:
            - Model Y: standard, premium, performance
            - Model 3: model_3, model_3_lr, model_3_performance
        range:
          type: string
          nullable: true
          enum: [maximale_reichweite, standard]
          description: Range option (not applicable for Performance)
        drive:
          type: string
          nullable: true
          enum: [rwd, awd]
          description: Drive type
        color:
          type: string
          nullable: true
          description: Exterior color
        interior:
          type: string
          nullable: true
          enum: [black, white]
          description: Interior color
        wheels:
          type: string
          nullable: true
          enum: ['18', '19', '20', '21']
          description: Wheel size
        towHitch:
          type: string
          nullable: true
          enum: [ja, nein, nv]
          description: Tow hitch option
        autopilot:
          type: string
          nullable: true
          enum: [none, ap, eap, eap_transfer, fsd, fsd_transfer]
          description: Autopilot package
        seats:
          type: string
          nullable: true
          enum: ['5', '7']
          description: Seat configuration (5 or 7 seater, 7 only for Model Y Premium AWD)
        source:
          type: string
          nullable: true
          enum: [tost]
          description: Order source — "tost" for TOST-managed orders, null for manually created orders
        tostUserId:
          type: string
          nullable: true
          description: TOST's internal user identifier (only set for TOST-managed orders)
        deliveryWindow:
          type: string
          nullable: true
          description: Expected delivery window
        deliveryLocation:
          type: string
          nullable: true
          description: Delivery location
        vin:
          type: string
          nullable: true
          description: Vehicle Identification Number
        vinReceivedDate:
          type: string
          nullable: true
          description: Date VIN was received (DD.MM.YYYY)
        papersReceivedDate:
          type: string
          nullable: true
          description: Date papers were received (DD.MM.YYYY)
        productionDate:
          type: string
          nullable: true
          description: Production date (DD.MM.YYYY)
        typeApproval:
          type: string
          nullable: true
          description: Type approval code
        typeVariant:
          type: string
          nullable: true
          description: Type variant code
        deliveryDate:
          type: string
          nullable: true
          description: Actual delivery date (DD.MM.YYYY)
        orderToProduction:
          type: integer
          nullable: true
          description: Days from order to production
        orderToVin:
          type: integer
          nullable: true
          description: Days from order to VIN received
        orderToDelivery:
          type: integer
          nullable: true
          description: Days from order to delivery
        orderToPapers:
          type: integer
          nullable: true
          description: Days from order to papers received
        papersToDelivery:
          type: integer
          nullable: true
          description: Days from papers to delivery
        archived:
          type: boolean
          description: Whether the order is archived
        archivedAt:
          type: string
          format: date-time
          nullable: true
          description: Archive timestamp
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
    CreateOrderRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Customer username (required)
        vehicleType:
          type: string
          enum: [Model Y, Model 3]
          description: Vehicle type (defaults to Model Y)
        orderDate:
          type: string
          description: Order date (DD.MM.YYYY format)
        country:
          type: string
        model:
          type: string
          description: |
            Model trim. Values depend on vehicleType:
            - Model Y: standard, premium, performance
            - Model 3: model_3, model_3_lr, model_3_performance
        range:
          type: string
          enum: [maximale_reichweite, standard]
        drive:
          type: string
          enum: [rwd, awd]
        color:
          type: string
        interior:
          type: string
          enum: [black, white]
        wheels:
          type: string
          enum: ['18', '19', '20', '21']
        towHitch:
          type: string
          enum: [ja, nein, nv]
        autopilot:
          type: string
          enum: [none, ap, eap, eap_transfer, fsd, fsd_transfer]
        seats:
          type: string
          enum: ['5', '7']
          description: Seat configuration (5 or 7 seater)
        deliveryWindow:
          type: string
        deliveryLocation:
          type: string
        vin:
          type: string
        vinReceivedDate:
          type: string
        papersReceivedDate:
          type: string
        productionDate:
          type: string
        typeApproval:
          type: string
        typeVariant:
          type: string
        deliveryDate:
          type: string
        editCode:
          type: string
          description: Custom edit password (min 6 chars, must contain a number)
...
    CreateOrderResponse:
      type: object
      properties:
        id:
          type: string
          description: Created order ID
        editCode:
          type: string
          nullable: true
          description: Edit code for future updates
        message:
          type: string
...
    Meta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        version:
          type: string
          example: "v1"
    Pagination:
      type: object
      properties:
        total:
          type: integer
          description: Total number of items
        limit:
          type: integer
          description: Items per page
        offset:
          type: integer
          description: Items skipped
        hasMore:
          type: boolean
          description: Whether more items exist
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        meta:
          $ref: '#/components/schemas/Meta'
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              description: Error code
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
              description: Field-specific error details
        meta:
          $ref: '#/components/schemas/Meta'
  responses:
    Unauthorized:
      description: API key missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "UNAUTHORIZED"
              message: "API key required. Use X-API-Key header."
            meta:
              timestamp: "2025-02-12T10:30:00.000Z"
              version: "v1"
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "VALIDATION_ERROR"
              message: "Validation failed"
              details:
                name: "Name is required"
            meta:
              timestamp: "2025-02-12T10:30:00.000Z"
              version: "v1"
    Conflict:
      description: Conflict (e.g., concurrent modification)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "CONFLICT"
              message: "Order was modified by another user. Please refresh and try again."
            meta:
              timestamp: "2025-02-12T10:30:00.000Z"
              version: "v1"
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "SERVER_ERROR"
              message: "Internal server error"
            meta:
              timestamp: "2025-02-12T10:30:00.000Z"
              version: "v1"